Implemented the basics of running the game environment, console player, etc. The game adds ConsolePlayers, asks for mulligans, can play lands, and can somehwat buggily play creatures.
Next on the todo list is to finish casting creatures, then add creature combat.
I need to consider how granular I want the actions to be - should tapping lands be part of casting a spell?  Do I want to ask what colors to use for casting the spell?
The code is really ugly right now.  Probably not the right time to be worrying about it.

+++ mlaSharp/Library.cs
@@ -36,8 +36,16 @@ namespace mlaSharp
 		/// <summary>
 		/// Draw a card.
 		/// </summary>
+		/// <exception cref="PlayerLostException">
+		/// Throws a <see cref="PlayerLostException"> if there are no cards in library and the player attempts to draw a card.
+		/// </exception>
 		public Card Draw()
 		{
+			if(this.Count == 0)
+			{
+				throw new PlayerLostException("Drawing a card with no cards left in library",null,Owner);
+			}
+			
 			Card item = this[0];
 			this.RemoveAt(0);	
 			return item;
@@ -49,8 +57,16 @@ namespace mlaSharp
 		/// <param name='x'>
 		/// The number of cards to draw
 		/// </param>
+		/// <exception cref="PlayerLostException">
+		/// Throws a <see cref="PlayerLostException"> if there are no cards in library and the player attempts to draw a card.
+		/// </exception>
 		public IEnumerable<Card> Draw(int x)
 		{
+			if(this.Count < x)
+			{
+				throw new PlayerLostException("Drawing a card with no cards left in library",null,Owner);	
+			}
+			
 			var cards = this.Take(x);
 			this.RemoveRange(0,x);
 			return cards;
